requirements.txt
Django==5.1.7
djangorestframework==3.15.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
django-filter==24.1 